+++ app/src/main/java/com/styudint/nepyatnashki/data/AndroidGameState.kt
@@ -1,0 +1,7 @@
+package com.styudint.nepyatnashki.data
+
+import androidx.lifecycle.LiveData
+
+interface AndroidGameState : GameState {
+    fun stopWatch(): LiveData<Long>
+}
