+++ app/src/main/java/com/styudint/nepyatnashki/data/GameStartStateGenerator.kt
@@ -3,5 +3,5 @@ package com.styudint.nepyatnashki.data
 import androidx.lifecycle.LiveData
 
 interface GameStartStateGenerator {
-    fun generate(): LiveData<GameState>
+    fun generate(): LiveData<AndroidGameState>
 }
